#57 Move submodules into core subpackage
