+++ pandas_dataclasses/__init__.py
@@ -1,11 +1,10 @@
+# type: ignore
+from . import core
+from .core.frame import *
+from .core.generic import *
+from .core.series import *
+from .core.typing import *
+
+
+# metadata
 __version__ = "0.3.0"
-
-
-# submodules
-from .parsers import *
-from .specs import *
-from .typing import *
-
-
-# subpackages
-from .pandas import *
