+++ pandas_dataclasses/core/__init__.py
@@ -1,0 +1,6 @@
+# type: ignore
+from . import frame
+from . import generic
+from . import specs
+from . import series
+from . import typing
