+++ pandas_dataclasses/core/generic.py
@@ -16,7 +16,7 @@ from typing_extensions import get_args, get_origin
 # submodules
 from .frame import asdataframe
 from .series import asseries
-from ..typing import AnyPandas, P, PandasClass
+from .typing import AnyPandas, P, PandasClass
 
 
 # type hints
