+++ pandas_dataclasses/core/series.py
@@ -13,8 +13,8 @@ from morecopy import copy
 
 
 # submodules
-from ..parsers import get_attrs, get_data, get_factory, get_index, get_name
-from ..typing import DataClass, P, PandasClass
+from .parsers import get_attrs, get_data, get_factory, get_index, get_name
+from .typing import DataClass, P, PandasClass
 
 
 # type hints
