--- pandas_dataclasses/pandas/__init__.py
@@ -1,4 +1,0 @@
-# submodules
-from .frame import *
-from .generic import *
-from .series import *
