+++ tests/test_frame.py
@@ -6,8 +6,8 @@ from typing import ClassVar
 # dependencies
 import numpy as np
 import pandas as pd
-from pandas_dataclasses.pandas.frame import AsDataFrame
-from pandas_dataclasses.typing import Attr, Data, Index
+from pandas_dataclasses.core.frame import AsDataFrame
+from pandas_dataclasses.core.typing import Attr, Data, Index
 from typing_extensions import Annotated as Ann
