+++ tests/test_series.py
@@ -6,8 +6,8 @@ from typing import ClassVar
 # dependencies
 import numpy as np
 import pandas as pd
-from pandas_dataclasses.pandas.series import AsSeries
-from pandas_dataclasses.typing import Attr, Data, Index, Name
+from pandas_dataclasses.core.series import AsSeries
+from pandas_dataclasses.core.typing import Attr, Data, Index, Name
 from typing_extensions import Annotated as Ann
 from typing_extensions import Literal as L
